pyproject.toml
[project]
name = "openfinance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.0.0",
    "pymupdf>=1.24.14",
]
